+++ src/config/keys.rs
@@ -158,6 +158,8 @@ fn default_binding(action: &Action) -> Option<(KeySpec, &'static str)> {
         Action::ToggleDiffStyle => (KeyCodeSpec::Char('s'), 4, "S"),
         Action::TreeCollapse => (KeyCodeSpec::Left, 0, "Left"),
         Action::TreeExpand => (KeyCodeSpec::Right, 0, "Right"),
+        Action::TreeCollapseAll => (KeyCodeSpec::Char('-'), 0, "-"),
+        Action::TreeExpandAll => (KeyCodeSpec::Char('='), 4, "+"),
         Action::Move(1) => (KeyCodeSpec::Char('j'), 0, "j"),
         Action::Move(-1) => (KeyCodeSpec::Char('k'), 0, "k"),
         Action::Page(1) => (KeyCodeSpec::Char('d'), 1, "Ctrl+d"),